Refactor: client, admin API 구분
