+++ src/client/src/api/client/location.ts
@@ -1,4 +1,4 @@
-import privateApi from "./private_api";
+import privateApi from "../private_api";
 
 export const getLocation = async() => {
     const response = await privateApi.get('/api/v1/location').catch((err) => {
